refactor : user Token 로직을 hook으로 변경
